EVM REPORT (document changes).md
A report on the documentation on Edgeware by DotMatrix

All seems great thus far.

Just a few things: the document UI of gitbook should be as appealing as the theme of the actual Edgeware website. Furthermore, add in a toggle button to switch between themes: light and dark mode. Dark mode is good for nocturnal coders:-). There are more small little issues.

1. Spelling error in line 3: change "effect" to "affect".
2. There is no link to the description on the PoS consensus. Added in this link [https://docs.edgewa.re/edgeware-runtime/consensus](https://docs.edgewa.re/edgeware-runtime/consensus) on consensus before the Aura page.